+++ doc/conf.py
@@ -22,7 +22,7 @@ copyright = '2015-2018, 2020-2023, Sean Whitton, 2021-2022 David Bremner'
 author = 'Sean Whitton'
 
 # The full version, including alpha/beta/rc tags
-release = '1.3.1'
+release = '1.3.2'
 
 
 # -- General configuration ---------------------------------------------------